# Modify implementation of tf.nn.batch_normalization