Give a slightly different implementation of tf.nn.batch_normalization so that per-example gradients needed by differentially private stochastic gradient descent can be calculated for gamma and beta in batch norm layers.