@rmlarsen 2017-08-09
@tensorflow-jenkins test this please