@liuzjmike 2017-08-24
Sorry for the late reply. I'm travelling right now. Will run the benchmark as soon as I get back.