@liuzjmike 2017-08-27
Following are the outputs of two consecutive run of batch_norm_benchmark.py on the original implementation and this new implementation, respectively. 

Original implementation:
```
> bazel test //tensorflow/python:batch_norm_benchmark
WARNING: Output base '/home/home5/zl99/.cache/bazel/_bazel_zl99/7128647531e7ce469ec086a63edb81ef' is on NFS. This may lead to surprising failures and undetermined behavior.
INFO: Found 1 test target...
Target //tensorflow/python:batch_norm_benchmark up-to-date:
  bazel-bin/tensorflow/python/batch_norm_benchmark
INFO: Elapsed time: 39.019s, Critical Path: 5.24s
//tensorflow/python:batch_norm_benchmark                                 PASSED in 5.1s

Executed 1 out of 1 test: 1 test passes.
There were tests whose specified size is too big. Use the --test_verbose_timeout_warnings command line option to see which ones these are.
```

New implementation:
```
> bazel test //tensorflow/python:batch_norm_benchmark
WARNING: Output base '/home/home5/zl99/.cache/bazel/_bazel_zl99/7128647531e7ce469ec086a63edb81ef' is on NFS. This may lead to surprising failures and undetermined behavior.
INFO: Found 1 test target...
Target //tensorflow/python:batch_norm_benchmark up-to-date:
  bazel-bin/tensorflow/python/batch_norm_benchmark
INFO: Elapsed time: 35.650s, Critical Path: 5.22s
//tensorflow/python:batch_norm_benchmark                                 PASSED in 5.1s

Executed 1 out of 1 test: 1 test passes.
There were tests whose specified size is too big. Use the --test_verbose_timeout_warnings command line option to see which ones these are.
```

I'm not sure if the benchmarks were run completely and correctly because of the last line of the output. But as far as the result shows, the two implementation should have similar performance.