@drpngx 2017-09-17
Sorry to be dense, but if it's the same speed and returns the same results, why change?