@liuzjmike 2017-09-17
This is not necessary an improvement to TensorFlow itself. But in order to apply differentially private stochastic gradient descent described in [this paper](https://arxiv.org/abs/1607.00133) by Abadi et al. on a network with batch normalization layers, per-example gradients need to be calculated for `gamma` and `beta`. The original implementation makes this calculation very hard in the TensorFlow framework while the new implementation significantly simplifies it.