@sguada 2017-09-18
You need to do blaze run not blaze test so it would print the actual timings.