@liuzjmike 2017-09-27
When I do bazel run I get this
```
> bazel run //tensorflow/python:batch_norm_benchmark
INFO: Found 1 target...
Target //tensorflow/python:batch_norm_benchmark up-to-date:
  bazel-bin/tensorflow/python/batch_norm_benchmark
INFO: Elapsed time: 389.044s, Critical Path: 289.23s

INFO: Running command line: bazel-bin/tensorflow/python/batch_norm_benchmark

----------------------------------------------------------------------
Ran 0 tests in 0.000s

OK
```
There are other jobs running on the cluster so reading the target was slow. But according to the output I think no test was run. Any idea why?

I will also try again when the cluster is less busy.