@drpngx 2017-09-27
You'll likely want to add the following to your command line:
```
-- --benchmarks=BatchNormBenchmark.benchmark_batch_norm
``