@liuzjmike 2017-09-27
```
bazel run //tensorflow/python:batch_norm_benchmark --benchmarks=BatchNormBenchmark.benchmark_batch_norm
```
gives `Unrecognized option: --benchmarks=BatchNormBenchmark.benchmark_batch_norm`.
```
bazel-bin/tensorflow/python/batch_norm_benchmark --benchmarks=BatchNormBenchmark.benchmark_batch_norm
```
gives the following output:

On original impl:
```
Forward convolution (lower layers).
2017-09-27 18:27:39.204075: W tensorflow/core/framework/op_def_util.cc:333] Op BatchNormWithGlobalNormalization is deprecated. It will cease to work in GraphDef version 9. Use tf.nn.batch_normalization().
cpu shape:4/3 #layers:10 mode:op scale:True train:False - 0.049103 secs
entry {
  name: "BatchNormBenchmark.batch_norm_cpu_input_shape_[8,128,128,32]_axes_[0,1,2]_mode_op_layers_10_scale_True_train_False"
  iters: 5
  wall_time: 0.0491034030914
}

cpu shape:4/3 #layers:10 mode:py scale:True train:False - 0.109263 secs
entry {
  name: "BatchNormBenchmark.batch_norm_cpu_input_shape_[8,128,128,32]_axes_[0,1,2]_mode_py_layers_10_scale_True_train_False"
  iters: 5
  wall_time: 0.109263372421
}

cpu shape:4/3 #layers:10 mode:slow scale:True train:False - 0.182931 secs
entry {
  name: "BatchNormBenchmark.batch_norm_cpu_input_shape_[8,128,128,32]_axes_[0,1,2]_mode_slow_layers_10_scale_True_train_False"
  iters: 5
  wall_time: 0.182931375504
}

=== op vs py: 122.5% ===
=== py vs slow: 67.4% ===
2017-09-27 18:27:42.277512: F tensorflow/core/common_runtime/gpu/gpu_device.cc:131] Check failed: err == cudaSuccess (2 vs. 0)
Abort (core dumped)
```

On proposed impl:
```
Forward convolution (lower layers).
2017-09-27 18:30:13.324701: W tensorflow/core/framework/op_def_util.cc:333] Op BatchNormWithGlobalNormalization is deprecated. It will cease to work in GraphDef version 9. Use tf.nn.batch_normalization().
cpu shape:4/3 #layers:10 mode:op scale:True train:False - 0.047454 secs
entry {
  name: "BatchNormBenchmark.batch_norm_cpu_input_shape_[8,128,128,32]_axes_[0,1,2]_mode_op_layers_10_scale_True_train_False"
  iters: 5
  wall_time: 0.0474540233612
}

cpu shape:4/3 #layers:10 mode:py scale:True train:False - 0.191871 secs
entry {
  name: "BatchNormBenchmark.batch_norm_cpu_input_shape_[8,128,128,32]_axes_[0,1,2]_mode_py_layers_10_scale_True_train_False"
  iters: 5
  wall_time: 0.191870594025
}

cpu shape:4/3 #layers:10 mode:slow scale:True train:False - 0.180249 secs
entry {
  name: "BatchNormBenchmark.batch_norm_cpu_input_shape_[8,128,128,32]_axes_[0,1,2]_mode_slow_layers_10_scale_True_train_False"
  iters: 5
  wall_time: 0.180248832703
}

=== op vs py: 304.3% ===
=== py vs slow: -6.1% ===
2017-09-27 18:30:33.873145: F tensorflow/core/common_runtime/gpu/gpu_device.cc:131] Check failed: err == cudaSuccess (2 vs. 0)
Abort (core dumped)