@drpngx 2017-09-27
Yes, if you use `bazel run` you need the extra leading `--` to indicate that it isn't a bazel option, but an option passed to the test.

The core dump is a bit worrisome, though. Could you look into it?