@drpngx 2017-09-27
Finally, the proposed implementation looks 2x slower?