@liuzjmike 2017-09-28
Thank you for the tip. I tried running with
```
bazel run //tensorflow/python:batch_norm_benchmark -- --benchmarks=BatchNormBenchmark.benchmark_batch_norm
```
and the core dump did not happen. It might be due to some mechanism of bazel.

Yes it looks like the proposed impl is sadly slower than the original one, maybe due to the original impl optimizing on the calculation dimension. Maybe we can give users an option in tf.contrib.layer.batch_norm to use the proposed impl if they want to do the per-example gradient operation. If you don't want to make that change, I'm ok with closing this PR.