@drpngx 2017-09-28
@sguada what's the right thing to do? Close or provide this implementation as an additional option?