@sguada 2017-10-14
Given the slowness I'm not inclined to include it, also adding extra options requires extra maintenance. So unless convinced otherwise I would close it.