@martinwicke 2017-11-07
Closing. It can always be reopened, or a new PR can be made.